Added new function to VisitService

+++ src/main/java/com/cgi/dentistapp/dto/VisitCreateDTO.java
@@ -16,7 +16,7 @@ import java.util.Date;
 @Setter
 public class VisitCreateDTO extends AbstractDTO {
 
-    private Integer visitId;
+    private Long visitId;
 
     @NotNull
     @DateTimeFormat(pattern = "dd.MM.yyyy")
